requirements.txt
flask==2.2.3
spotipy==2.19.0
pandas==1.4.2
gunicorn
numpy==1.22
Pillow==9.1.1
tensorflow==2.12.0
keras_preprocessing==1.1.2
keras==2.12.0
opencv_python_headless==4.5.5.64
urllib3==1.26.12